updated where I'm storing data
