+++ src/js/data/siteList.js
@@ -1,0 +1,85 @@
+const siteList = [
+    {
+        name: 'AimHigher Web Design',
+        slug: 'aimhigher',
+        url: 'aimhigherwebdesign.com.au',
+        date: 'November 2017',
+        mobile: true,
+        current: true,
+        github: 'https://github.com/AimHigher-Web-Design/aimhigher',
+    },
+    {
+        name: 'Treasure Pics',
+        slug: 'treasure-pics',
+        url: 'www.treasurepics.com.au/',
+        date: 'January 2015',
+        mobile: false,
+        current: true,
+        github: false,
+    },
+    {
+        name: 'Amy Goes to Perth',
+        slug: 'amy-goes-to-perth',
+        url: 'amygoestoperth.com.au/',
+        date: 'October 2017',
+        mobile: true,
+        current: true,
+        github: 'https://github.com/amykapernick/amygoestoperth',
+    },
+    {
+        name: 'Nadine Smith Studio',
+        slug: 'nadine-smith',
+        url: 'nadinesmithstudio.com.au',
+        date: 'June 2017',
+        mobile: true,
+        current: false,
+        github: 'https://github.com/amykapernick/nadinesmithstudio',
+    },
+    {
+        name: 'Eco Spray Tans Perth',
+        slug: 'eco-spray-tans',
+        url: 'www.ecospraytansperth.com.au/',
+        date: 'October 2015',
+        mobile: true,
+        current: true,
+        github: false,
+    },
+    {
+        name: 'Glenrock Hay',
+        slug: 'glenrock-hay',
+        url: 'www.glenrockhay.com',
+        date: 'December 2013 and December 2014',
+        mobile: true,
+        current: true,
+        github: false,
+    },  
+    {
+        name: 'Covisor',
+        slug: 'covisor',
+        url: 'www.covisor.com.au/',
+        date: 'January 2015',
+        mobile: true,
+        current: true,
+        github: false,
+    },
+    {
+        name: 'Collards Fuel Supplies',
+        slug: 'collards-fuel',
+        url: 'www.collardsfuelsupplies.com.au/',
+        date: 'January 2015',
+        mobile: true,
+        current: true,
+        github: false,
+    },
+    {
+        name: "Queensland Rural, Regional & Remote Women's Network",
+        slug: 'qrrrwn',
+        url: 'qrrrwn.org.au',
+        date: 'February 2015',
+        mobile: false,
+        current: false,
+        github: false,
+    },
+];
+
+export default siteList;
